+++ doc/CHANGELOG.rst
@@ -8,8 +8,13 @@ Changed
 ^^^^^^^
 * Revise Makefile.
 * Improve documentation.
-* Improve positioning info for SMILES integer-too-large errors
+* Improve positioning info for SMILES integer-too-large errors.
 * Allow parsing of deeply-nested SMILES.
+
+Fixed
+^^^^^
+* SMILES: Fix implicit bond order between aromatic atoms.
+
 
 `v0.2`_ - 2017-07-21
 --------------------